Move `NoteDao` and `NoteEntity` to  each package
