+++ app/src/main/java/com/laam/laamnotes/framework/datasource/NoteDataSource.kt
@@ -2,8 +2,8 @@ package com.laam.laamnotes.framework.datasource
 
 import com.laam.core.data.Note
 import com.laam.core.repository.NoteDataSource
-import com.laam.laamnotes.framework.db.NoteDao
-import com.laam.laamnotes.framework.db.NoteEntity
+import com.laam.laamnotes.framework.db.dao.NoteDao
+import com.laam.laamnotes.framework.db.entity.NoteEntity
 import javax.inject.Inject
 
 class NoteDataSource @Inject constructor(private val noteDao: NoteDao) :
