+++ app/src/main/java/com/laam/laamnotes/framework/db/AppDatabase.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.laam.laamnotes.framework.db.dao.NoteDao
+import com.laam.laamnotes.framework.db.entity.NoteEntity
 
 @Database(entities = [NoteEntity::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
