+++ app/src/main/java/com/laam/laamnotes/framework/db/entity/NoteEntity.kt
@@ -1,4 +1,4 @@
-package com.laam.laamnotes.framework.db
+package com.laam.laamnotes.framework.db.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
